演讲比赛管理系统，优化界面显示，优化csv文件存入和读取

+++ Speech_Contest_Management_System/Speech_Contest_Management_System/participant.h
@@ -3,7 +3,7 @@
 class Participant {
 public:
 	string m_Id; //编号
-	int m_point; //平均分
+	double m_point; //平均分
 
 public:
